+++ core-tests/jvm/src/test/scala/zio/CancelableFutureSpecJVM.scala
@@ -1,6 +1,5 @@
 package zio
 
-import zio.internal.Executor
 import zio.test.Assertion._
 import zio.test.TestAspect._
 import zio.test._
