+++ core/js/src/main/scala/zio/ExecutorPlatformSpecific.scala
@@ -14,6 +14,6 @@
  * limitations under the License.
  */
 
-package zio.internal
+package zio
 
 trait ExecutorPlatformSpecific
