+++ core/jvm/src/main/scala/zio/internal/Blocking.scala
@@ -1,5 +1,7 @@
 package zio.internal
 
+import zio.Executor
+
 import java.util.concurrent.{SynchronousQueue, ThreadPoolExecutor, TimeUnit}
 
 object Blocking {
