+++ core/jvm/src/main/scala/zio/internal/ZScheduler.scala
@@ -16,6 +16,8 @@
 
 package zio.internal
 
+import zio.Executor
+
 import java.util.concurrent.ThreadLocalRandom
 import java.util.concurrent.atomic.AtomicInteger
 import java.util.concurrent.locks.LockSupport
