+++ core/shared/src/main/scala/zio/RIO.scala
@@ -16,8 +16,6 @@
 
 package zio
 
-import zio.internal.Executor
-
 import scala.concurrent.ExecutionContext
 import scala.reflect.ClassTag
