+++ core/shared/src/main/scala/zio/ZManaged.scala
@@ -17,7 +17,6 @@
 package zio
 
 import zio.ZManaged.ReleaseMap
-import zio.internal.Executor
 
 import scala.collection.immutable.LongMap
 import scala.concurrent.ExecutionContext
