+++ test/shared/src/main/scala/zio/test/mock/Mock.scala
@@ -16,10 +16,9 @@
 
 package zio.test.mock
 
-import zio.internal.Executor
 import zio.stream.{ZSink, ZStream}
 import zio.test.TestPlatform
-import zio.{Has, Runtime, Tag, ULayer, URIO, URLayer, ZIO}
+import zio.{Executor, Has, Runtime, Tag, ULayer, URIO, URLayer, ZIO}
 
 /**
  * A `Mock[R]` represents a mockable environment `R`.
